Fix uint8 mask-bit overflow in downsample binning and saturation removal
_aggregate_sorted_groups did (mask_grouped & ignore_mask) where
ignore_mask can be 4096 (bit 12, PS1 saturation, the ignore_mask_bits
default). PS1 mask arrays are meant to always be uint16 (the convolved
store writer normalizes on write), but this SCC's on-disk convolved.zarr
has 15 stale uint8 mask arrays predating that normalization; numpy's
strict same-dtype casting raises OverflowError on the bitwise AND instead
of the old silent wraparound. Upcast mask_grouped to int64 before the AND
so a narrower on-disk dtype degrades to "nothing ignored via that bit"
instead of crashing the whole downsample stage.

band_utils.py's remove_background has the identical unguarded pattern
(mask & 0x1000) in its quality-flag removal pass; it's wrapped in a
broad except-and-warn so it wasn't stage-fatal, but the same crash was
silently swallowing the quality-flag pass. Fixed with the same int64
upcast for when ps1_process is next run on this data.

+++ syndiff_pipeline/template_creation/processing/downsample.py
@@ -569,7 +569,10 @@ def _aggregate_sorted_groups(
     ps1_grouped = ps1_rav[slice_start:]
     mask_grouped = ps1_mask_rav[slice_start:]
 
-    ignored = (mask_grouped & ignore_mask) > 0
+    # mask dtype on disk may be narrower (e.g. uint8) than the ignore_mask bit
+    # position being tested (e.g. bit 12 = 4096); upcast before the bitwise AND
+    # so numpy's strict same-dtype casting doesn't reject the Python int.
+    ignored = (mask_grouped.astype(np.int64, copy=False) & int(ignore_mask)) > 0
     sum_weights = np.where(ignored, 0.0, ps1_grouped).astype(np.float64, copy=False)
     sum_weights = np.where(np.isnan(sum_weights), 0.0, sum_weights)
